requirements.txt
# Python package requirements for the multi-agent report writer example
langgraph
langchain-core
pydantic
google-generativeai
tavily
# Optional / recommended
python-dotenv

# Resilience & Testing
tenacity
pytest-cov